fix: resolve 14 critical payment/wallet bugs

+++ app/booking/models.py
@@ -9,7 +9,7 @@ This module defines the database models for booking operations:
 - Payment and transaction models
 """
 
-from sqlalchemy import Column, Integer, String, DateTime, Boolean, Text, Float, JSON, ForeignKey, Enum
+from sqlalchemy import Column, Integer, String, DateTime, Boolean, Text, Numeric, JSON, ForeignKey, Enum
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 import enum
@@ -150,9 +150,9 @@ class FlightBooking(Base, TenantMixin):
     passenger_details = Column(JSON, nullable=False)  # Store passenger info as JSON
     
     # Pricing and payment
-    base_fare = Column(Float, nullable=False)
-    taxes = Column(Float, nullable=False)
-    total_amount = Column(Float, nullable=False)
+    base_fare = Column(Numeric(14, 2), nullable=False)
+    taxes = Column(Numeric(14, 2), nullable=False)
+    total_amount = Column(Numeric(14, 2), nullable=False)
     currency = Column(String(3), default="INR")
     payment_method = Column(Enum(PaymentMethod), nullable=False)
     payment_status = Column(Enum(PaymentStatus), default=PaymentStatus.PENDING)
@@ -164,7 +164,7 @@ class FlightBooking(Base, TenantMixin):
     # Additional information
     special_requests = Column(Text, nullable=True)
     cancellation_reason = Column(Text, nullable=True)
-    refund_amount = Column(Float, nullable=True)
+    refund_amount = Column(Numeric(14, 2), nullable=True)
     
     # Audit fields
     created_at = Column(DateTime(timezone=True), server_default=func.now())
@@ -215,10 +215,10 @@ class HotelBooking(Base, TenantMixin):
     special_requests = Column(Text, nullable=True)
     
     # Pricing and payment
-    room_rate = Column(Float, nullable=False)  # Rate per night
-    base_amount = Column(Float, nullable=False, default=0.0)
-    taxes = Column(Float, nullable=False, default=0.0)
-    total_amount = Column(Float, nullable=False)
+    room_rate = Column(Numeric(14, 2), nullable=False)  # Rate per night
+    base_amount = Column(Numeric(14, 2), nullable=False, default=0.0)
+    taxes = Column(Numeric(14, 2), nullable=False, default=0.0)
+    total_amount = Column(Numeric(14, 2), nullable=False)
     currency = Column(String(3), default="INR")
     payment_method = Column(Enum(PaymentMethod), nullable=False)
     payment_status = Column(Enum(PaymentStatus), default=PaymentStatus.PENDING)
@@ -229,7 +229,7 @@ class HotelBooking(Base, TenantMixin):
     
     # Additional information
     cancellation_reason = Column(Text, nullable=True)
-    refund_amount = Column(Float, nullable=True)
+    refund_amount = Column(Numeric(14, 2), nullable=True)
     cancellation_policy = Column(String(100), nullable=True)
     
     # Audit fields
@@ -277,10 +277,10 @@ class BusBooking(Base, TenantMixin):
     passenger_details = Column(JSON, nullable=False)  # Store passenger info as JSON
     
     # Pricing and payment
-    fare_per_passenger = Column(Float, nullable=False)
-    base_amount = Column(Float, nullable=False, default=0.0)
-    taxes = Column(Float, nullable=False, default=0.0)
-    total_amount = Column(Float, nullable=False)
+    fare_per_passenger = Column(Numeric(14, 2), nullable=False)
+    base_amount = Column(Numeric(14, 2), nullable=False, default=0.0)
+    taxes = Column(Numeric(14, 2), nullable=False, default=0.0)
+    total_amount = Column(Numeric(14, 2), nullable=False)
     currency = Column(String(3), default="INR")
     payment_method = Column(Enum(PaymentMethod), nullable=False)
     payment_status = Column(Enum(PaymentStatus), default=PaymentStatus.PENDING)
@@ -294,7 +294,7 @@ class BusBooking(Base, TenantMixin):
     dropping_point = Column(String(255), nullable=True)
     special_requests = Column(Text, nullable=True)
     cancellation_reason = Column(Text, nullable=True)
-    refund_amount = Column(Float, nullable=True)
+    refund_amount = Column(Numeric(14, 2), nullable=True)
     
     # Audit fields
     created_at = Column(DateTime(timezone=True), server_default=func.now())
